feat: add image generation tool
