+++ src/infrastructure/tools/current_stock_price.py
@@ -1,0 +1,35 @@
+from typing import Type
+
+import yfinance as yf
+from langchain.tools import BaseTool
+from pydantic import BaseModel, Field
+
+
+def get_current_stock_price(ticker):
+    """Method to get current stock price"""
+
+    ticker_data = yf.Ticker(ticker)
+    recent = ticker_data.history(period="1d")
+    return {"price": recent.iloc[0]["Close"], "currency": ticker_data.info["currency"]}
+
+
+class CurrentStockPriceInput(BaseModel):
+    """Inputs for get_current_stock_price"""
+
+    ticker: str = Field(description="Ticker symbol of the stock")
+
+
+class CurrentStockPriceTool(BaseTool):
+    name = "get_current_stock_price"
+    description = """
+        Useful when you want to get current stock price.
+        You should enter the stock ticker symbol recognized by the yahoo finance
+        """
+    args_schema: Type[BaseModel] = CurrentStockPriceInput
+
+    def _run(self, ticker: str):
+        price_response = get_current_stock_price(ticker)
+        return price_response
+
+    def _arun(self, ticker: str):
+        raise NotImplementedError("get_current_stock_price does not support async")
