+++ src/infrastructure/tools/stock_performance.py
@@ -6,14 +6,6 @@ from langchain.tools import BaseTool
 from pydantic import BaseModel, Field
 
 
-def get_current_stock_price(ticker):
-    """Method to get current stock price"""
-
-    ticker_data = yf.Ticker(ticker)
-    recent = ticker_data.history(period="1d")
-    return {"price": recent.iloc[0]["Close"], "currency": ticker_data.info["currency"]}
-
-
 def get_stock_performance(ticker, days):
     """Method to get stock price change in percentage"""
 
@@ -23,28 +15,6 @@ def get_stock_performance(ticker, days):
     old_price = history.iloc[0]["Close"]
     current_price = history.iloc[-1]["Close"]
     return {"percent_change": ((current_price - old_price) / old_price) * 100}
-
-
-class CurrentStockPriceInput(BaseModel):
-    """Inputs for get_current_stock_price"""
-
-    ticker: str = Field(description="Ticker symbol of the stock")
-
-
-class CurrentStockPriceTool(BaseTool):
-    name = "get_current_stock_price"
-    description = """
-        Useful when you want to get current stock price.
-        You should enter the stock ticker symbol recognized by the yahoo finance
-        """
-    args_schema: Type[BaseModel] = CurrentStockPriceInput
-
-    def _run(self, ticker: str):
-        price_response = get_current_stock_price(ticker)
-        return price_response
-
-    def _arun(self, ticker: str):
-        raise NotImplementedError("get_current_stock_price does not support async")
 
 
 class StockPercentChangeInput(BaseModel):
